+++ tests/Integration/parsers/LineParserTest.php
@@ -14,6 +14,12 @@ use ValueParsers\ValueParser;
  */
 class LineParserTest extends \ValueParsers\Test\StringValueParserTest {
 
+	public function setUp() {
+		if ( !defined( 'MEDIAWIKI' ) ) {
+			$this->markTestSkipped( 'MediaWiki is not available' );
+		}
+	}
+
 	/**
 	 * @return string
 	 */